script get_match
improving get_dataset.py to get_matchs_dataset.py, it can take all the datas for some years and put it into the form of a csv file.
nameToID is a csv for mapping team's name on the flashscore site to our ids.
Match.csv -> the dataset for the year 2021-2022